Fix missing laser energy bars (add new DC fields to contain the bars).
DTM and others use the inside of the laser energy bars for DC displays. However,
the location of these HUD elements changes ever so slightly depending on the
in-game resolution and sometimes they just disappear because of this. To fix
this problem, I'm introducing a few new DC source elements that track the inside
of the laser energy bars. That way, it's easier to properly set UVs that will
work on all resolutions. These are the new fields:

       DUAL_LASER_ENERGY_BAR_L_SRC
       DUAL_LASER_ENERGY_BAR_R_SRC
       QUAD_LASER_ENERGY_BAR_LU_SRC
       QUAD_LASER_ENERGY_BAR_LD_SRC
       QUAD_LASER_ENERGY_BAR_RU_SRC
       QUAD_LASER_ENERGY_BAR_RD_SRC
       ALL_LASER_ENERGY_BAR_SRC