Refactor DcEnergyBar logic, capture ions and lasers.
Ions and Lasers are rendered with the same enclosing box, but different inner
textures, so there's two (2) separate draw calls. Because of that, we need to
accumulate all boxes seen in one frame, and then at the end of the frame we can
compute the DC source elements for the ions and lasers bars.

The logic is split between the accumulation of boxes during regular rendering,
in Direct3DDevice::Execute(); and the final shrinking and labeling of the boxes
at the end of the frame, in PrimarySurface::Flip().