Disable Raytracing when an OPT is loaded in the Skirmish selection screen.
The new hook_concourse can load an OPT to make a quick render on the Skirmish
selection screen. This will trigger the multi-blas BVH builder, though, and then
Hilarity Ensues because this is not supposed to happen on a 2D screen.

I'm disabling a number of paths that will trigger BVH builds by also checking
that we are, in fact, rendering 3D content during regular flight.