Display the Skirmish OPT in VR.
Use RenderSkirmishOPT() to render the ship in the skirmish selection screen. We
need to add a new path to render just this case because we use array textures in
VR mode and _offscreenBufferHd is not an array texture. Consequently, we also
need to add _depthStencilHd and _depthStencilViewHd (non-array depth stencil) to
properly render the depth of the Skirmish OPT.