Fix Enhanced HUD text spacing in VR.
For certain in-game resolutions (1024x768) and font sizes, the spacing for text
for the enhanced HUD sometimes is wrong, causing text to encroach on itself. To
fix this, we must apply the scale factors we use in pancake mode -- but not the
displacement. We need to capture these scale factors and also apply them to the
vrCenterX to keep the text centered.