Fix the scale of the Enhanced HUD vertical bars in VR.
By using the size of the main target bracket, I can compute a scale that makes
the vertical bars shorter or longer as needed. I just eyeballed the values and
they seem to work.

Also fixed the transform chain (it was causing distortion when the brackets
neared the edge of the screen).