Fix Skirmish OPT rendering for high SteamVR resolutions.
High (150%, 2.5K pixels) SteamVR resolutions caused the OPT to look way too big,
or to just clip outside the view. To fix this, I'm scaling the OPT and moving it
so that it's more-or-less where it should be. This is a bit of a hack, but it
works, so moving on for now...