Fix the VR Gloves "disappearing" when the S-Foils are activated in the B/W.
It just so happens that we capture the Cockpit Constants for the first mesh we
see while rendering the cockpit, and we use these constants to place the VR
gloves. If said mesh is part of an S-Foils system, then the captured constants
will apply the S-Foils rotation to the VR Gloves and Hilarity Ensues.

To prevent this, I'm adding a new attribute to the .ac files:

	vr_anchor = 1

We just need to add this attribute to any texture with an AC action. That way,
we know which mesh is safe to capture for the VR Gloves.