Enable Triangle Pointer render for pancake mode.
The triangle pointer is already rendered around the center of the HMD when VR is
activated. This is necessary in VR because the edges of the screen are not
visible. However, it's also a good idea to do this in pancake mode because that
makes it easier to locate the target.

This feature is currently activated by default. To disable it, add:

	triangle_pointer_enabled = 0

To SSAO.cfg