Enable custom CubeMaps (WIP).
Here we're piggy-backing on top of the existing code that renders
DefaultStarfield.dds to render custom cubemaps that we can load from disk. The
array used to build the cubemap has the following convention for each index:

	0: Right
	1: Left
	2: Top
	3: Bottom
	4: Front
	5: Back

The cubemap is loaded from Effects\CubeMaps. Each file must be a 1024x1024 pix
image and should be numbered with the scheme above. PNG files are supported.