Use mission .ini files to load cubemaps.
Cubemaps can now be speficied on a per-mission basis. Just add:

	[CubeMaps]
	AllRegions = Effects\CubeMaps\Cube1\HiRes,3840

Where the integer after the comma is the resolution of the cubemap (this should
be removed later).

For skirmish missions, add a file called Skirmish\temp.ini and add the [CubeMaps]
tag there.

Also addded a "master switch" for cubemaps in SSAO.cfg:

	enable_cubemaps = 1

The default value is 0.