Enable multi-region cubemaps, use "[SkyBoxes]".
Changed [CubeMaps] for [SkyBoxes] because that's what most people use out there.

The last digit in the filename is now used to sort the cubemaps so that they
display correctly.

Per-region cubemaps are now enabled. Regions start at 0, and the max region is
3. "AllRegions" specifies a cubemap to use for all the regions in a mission.
However, specific regions can override AllRegions, example:

	AllRegions = Effects\CubeMaps\Cube2
	Region1 = Effects\CubeMaps\Cube0
	Region2 = Effects\CubeMaps\Cube3

Regions 0 and 3 will display the map specified in AllRegions.