Enable CubeMap rotation.
Basic X,Y,Z rotation of the cubemaps is now possible. Use the following tags
in the .ini file:

	AllRegionsRotationX
	AllRegionsRotationY
	AllRegionsRotationZ

	RegionNRotationX
	RegionNRotationY
	RegionNRotationZ

For N = 0..3. All angles are in degrees and all rotations are applied along the
global (World) X,Y,Z axes.

Also, to reload the settings in the .ini file, press Ctrl+Alt+L, and enable
Developer Mode in SSAO.cfg.