Enable lo-res illumination CubeMaps.
Instead of sampling the hi-res CubeMap and using it to tint the Sun's color,
let's (optionally) load a lo-res CubeMap and use that instead. The lo-res
version isn't displayed and it's only used to tint the color of the
illumination. That way we can control that separately. New tags added:

	AllRegionsIllum = <path>
	RegionNIllum = <path>

For N in 0..3