Fix semi-transparent backdrop planets.
The default blending mode for backdrops is a bit weird: if a semi-transparent
nebula is in front of a solid planet, then the planet will also become
semi-transparent. To fix this, I've added a custom blend mode for backdrops
that accumulates the alpha channel. This seems to work, but I also had to
brighten the DefaultStarfield because now it was way too faint.