Enable overlay CubeMaps.
A second cubemap layer can now be displayed. If the standard skybox is located
in <path> then the overlay must be located in <path>_overlay. Otherwise, the
following tags can be used to select the path for the overlay cubemap:

	AllRegionsOverlay
	RegionNOverlay (for N = 0..3)

The overlay cubemap is rendered after the backdrops and alpha blending is
supported.