Add support for CubeMap rotation editing during flight (WIP).
It is now possible to edit the rotation of a CubeMap while flying. To enable
this add:

	debug_cubemap_rotation = {standard|local_coords}

to SSAO.cfg, then use Ctrl+Arrow Keys to change the rotation of the cubemap, and
and Ctrl+Shift+Left/Right Arrow Keys to change the CubeMap's spin. It is highly
recommended that a cross is added to the center of the front (4) side of the
CubeMap or else the rotation may be difficult to understand.

The "standard" mode applies azimuth-elevation-spin rotation to the CubeMap (this
is equivalent to the current rotation angles).

The "local_coords" mode moves the CubeMap relative to its current orientation
(this makes more sense when the cross is added to the front side of the cube).

In both modes, the settings are autosaved to the current .ini file. In the
"local_coords" case, new tags are used to store the orientation of the CubeMap:

	AllRegionsRVector = Vector3
	AllRegionsUVector = Vector3
	AllRegionsFVector = Vector3

	RegionNRVector = Vector3
	RegionNUVector = Vector3
	RegionNFVector = Vector3 (for N = 0..3)

Only one mode is supported per region. local_coords settings are currently not
read from the file, and only the AllRegions settings are saved for now (that's
why I put "WIP" in the summary line above!)