Sideload CubeMaps through XwaDDrawPlayer (courtesy of Jeremy Ansel)
But only if XwaDDrawPlayer.exe is run before XWA. I disabled the code that
Jeremy added to run XwaDDrawPlayer.exe because it seems to crash when loading
the first CubeMap. So, if the helper process isn't running, then we just revert
to loading the CubeMaps directly from the FX ddraw.

I also added this setting to SSAO.cfg:

	enable_xwaddrawplayer = 1

To force-disable this feature if necessary.