Force 60 FPS by default.
Lots of people get into trouble because they play the game at more than 60 FPS.
In order to avoid this, Jeremy suggested this change to force 60 FPS by default.
I've added a "chicken-switch" to SSAO.cfg to turn this off if it causes more
problems:

	force_60_fps (default is 1)

Just set it to 0 to disable this feature and restart the game.