Load proper edit angles for "standard" mode when reloading mission.
Set:

debug_cubemap_rotation = standard

in SSAO.cfg, then load a mission with nonzero AllRegions CubeMap rotation, then
exit the mission and go back to the mission selection screen. Load the same
mission again, and the rotation will be reset back to 0, ignoring the CubeMap
rotation in the .ini file. This happened because the system was failing to
detect that a new mission was loaded and the edit angles were set to 0, instead
of getting the initial values from the .ini file. The problem only happens in
"edit mode" for the "standard mode" (the "local_coords" mode always resets the
edit angles to 0).

To fix this, I added a new function that returns 0xFFFF when no CubeMap is
active in the current region so that -1 can be used for the AllRegions CubeMap
and the change can be easily detected.